Fix bug where all things go to APRS-IS.
A bug was introduced that prevented strings from being displayed on
the screen.  This happened because the $PSParameterSetName was not
carried into the hidden function Send-APRSThing.  This has been
corrected by completely removing the Server, Port, and Force para-
meters when they are not included.  (I suppose $null wasn't good
enough!)