Add support for position reports.
This commit and release includes support for attaching a latitude
and longitude to messages, announcements, bulletins, and group
bulletins.  This will allow compatible receivers (APRS.fi included)
to plot them on a map.

Due to limitations in how I coded this, you cannot include a position
report if you are exactly at 0,0.  Change it to 0.0001,0 if you happen
to be floating there and using this app.  (Alternative implementations
are welcome!)